backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
boto3==1.34.23
openai==1.10.0
langchain==0.1.1
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2
PyMuPDF==1.23.15
pytesseract==0.3.10
python-dotenv==1.0.1
httpx==0.26.0
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
